+++ micro-ui/web/micro-ui-internals/packages/svg-components/src/index.js
@@ -815,8 +815,8 @@ import { SetUpMicroplanSvg } from "./svg/SetUpMicroplanSvg";
 import { AddCircle } from "./svg/AddCircle";
 import { GlobeLocationPin } from "./svg/GlobeLocationPin";
 import { NewWindow } from "./svg/NewWindow";
-import { Outpatient_med } from "./svg/Outpatient_med";
-import { Ad_units } from "./svg/Ad_units";
+import { OutpatientMed } from "./svg/OutpatientMed";
+import { AdUnits } from "./svg/AdUnits";
 
 export {
   AddCircle,
@@ -1636,6 +1636,6 @@ export {
   MeasurementMuktaIcon,
   GlobeLocationPin,
   NewWindow,
-  Outpatient_med,
-  Ad_units
+  OutpatientMed,
+  AdUnits
 };